[gpu-video] Make `VideoDevice` backend agnostic
