+++ gpu-video/src/backends/vulkan/codec/h265.rs
@@ -3,51 +3,59 @@ use std::ptr::null_mut;
 use ash::vk;
 
 use crate::{
-    codec::{
+    backends::vulkan::codec::{
         Codec, CodecCapabilities, CodecSpecificDecodeCapabilities, CodecSpecificEncodeCapabilities,
         CodecSpecificEncoderQualityLevelProperties,
-        h264::parameters::{VkH264PictureParameterSet, VkH264SequenceParameterSet},
+        h265::parameters::{
+            VkH265PictureParameterSet, VkH265SequenceParameterSet, VkH265VideoParameterSet,
+        },
     },
-    device::caps::NativeEncodeH264Capabilities,
-    parameters::H264Profile,
+    backends::vulkan::vulkan_device::caps::{
+        NativeEncodeCapabilities, NativeEncodeH265Capabilities,
+    },
+    parameters::H265Profile,
 };
 
 pub(crate) mod encode;
 pub(crate) mod parameters;
 
+pub(crate) struct H265CodecParameters {
+    pub(crate) vps: Vec<VkH265VideoParameterSet>,
+    pub(crate) sps: Vec<VkH265SequenceParameterSet>,
+    pub(crate) pps: Vec<VkH265PictureParameterSet>,
+}
+
+pub(crate) struct H265VkParameters {
+    pub(crate) vps: Vec<vk::native::StdVideoH265VideoParameterSet>,
+    pub(crate) sps: Vec<vk::native::StdVideoH265SequenceParameterSet>,
+    pub(crate) pps: Vec<vk::native::StdVideoH265PictureParameterSet>,
+}
+
 #[derive(Debug, Clone, Copy)]
-pub(crate) struct H264Codec;
+pub(crate) struct H265Codec;
 
-pub(crate) struct H264CodecParameters {
-    pub(crate) sps: Vec<VkH264SequenceParameterSet>,
-    pub(crate) pps: Vec<VkH264PictureParameterSet>,
-}
+impl Codec for H265Codec {
+    type Profile = H265Profile;
 
-pub(crate) struct H264VkParameters {
-    pub(crate) sps: Vec<vk::native::StdVideoH264SequenceParameterSet>,
-    pub(crate) pps: Vec<vk::native::StdVideoH264PictureParameterSet>,
-}
+    type OwnedParameters = H265CodecParameters;
 
-impl Codec for H264Codec {
-    type Profile = H264Profile;
-
-    type OwnedParameters = H264CodecParameters;
-    type VkParameters<'a> = H264VkParameters;
+    type VkParameters<'a> = H265VkParameters;
 
     type VideoDecodeSessionParametersAddInfo<'a> =
-        vk::VideoDecodeH264SessionParametersAddInfoKHR<'a>;
+        vk::VideoDecodeH265SessionParametersAddInfoKHR<'a>;
     type VideoDecodeSessionParametersCreateInfo<'a> =
-        vk::VideoDecodeH264SessionParametersCreateInfoKHR<'a>;
+        vk::VideoDecodeH265SessionParametersCreateInfoKHR<'a>;
 
     type VideoEncodeSessionParametersAddInfo<'a> =
-        vk::VideoEncodeH264SessionParametersAddInfoKHR<'a>;
+        vk::VideoEncodeH265SessionParametersAddInfoKHR<'a>;
     type VideoEncodeSessionParametersCreateInfo<'a> =
-        vk::VideoEncodeH264SessionParametersCreateInfoKHR<'a>;
+        vk::VideoEncodeH265SessionParametersCreateInfoKHR<'a>;
 
     fn decode_parameters_add_info<'a: 'b, 'b>(
         parameters: &'b Self::VkParameters<'a>,
     ) -> Self::VideoDecodeSessionParametersAddInfo<'b> {
-        vk::VideoDecodeH264SessionParametersAddInfoKHR::default()
+        vk::VideoDecodeH265SessionParametersAddInfoKHR::default()
+            .std_vp_ss(&parameters.vps)
             .std_sp_ss(&parameters.sps)
             .std_pp_ss(&parameters.pps)
     }
@@ -55,7 +63,8 @@ impl Codec for H264Codec {
     fn decode_parameters_create_info<'a: 'b, 'b>(
         add_info: &'b Self::VideoDecodeSessionParametersAddInfo<'a>,
     ) -> Self::VideoDecodeSessionParametersCreateInfo<'b> {
-        vk::VideoDecodeH264SessionParametersCreateInfoKHR::default()
+        vk::VideoDecodeH265SessionParametersCreateInfoKHR::default()
+            .max_std_vps_count(32)
             .max_std_sps_count(32)
             .max_std_pps_count(32)
             .parameters_add_info(add_info)
@@ -64,7 +73,8 @@ impl Codec for H264Codec {
     fn encode_parameters_add_info<'a: 'b, 'b>(
         parameters: &'b Self::VkParameters<'a>,
     ) -> Self::VideoEncodeSessionParametersAddInfo<'b> {
-        vk::VideoEncodeH264SessionParametersAddInfoKHR::default()
+        vk::VideoEncodeH265SessionParametersAddInfoKHR::default()
+            .std_vp_ss(&parameters.vps)
             .std_sp_ss(&parameters.sps)
             .std_pp_ss(&parameters.pps)
     }
@@ -72,24 +82,25 @@ impl Codec for H264Codec {
     fn encode_parameters_create_info<'a: 'b, 'b>(
         add_info: &'b Self::VideoEncodeSessionParametersAddInfo<'a>,
     ) -> Self::VideoEncodeSessionParametersCreateInfo<'b> {
-        vk::VideoEncodeH264SessionParametersCreateInfoKHR::default()
+        vk::VideoEncodeH265SessionParametersCreateInfoKHR::default()
+            .max_std_vps_count(32)
             .max_std_sps_count(32)
             .max_std_pps_count(32)
             .parameters_add_info(add_info)
     }
 }
 
-impl CodecCapabilities for H264Codec {
-    type CodecSpecificDecodeCapabilities<'a> = vk::VideoDecodeH264CapabilitiesKHR<'a>;
-    type CodecSpecificEncodeCapabilities<'a> = vk::VideoEncodeH264CapabilitiesKHR<'a>;
+impl CodecCapabilities for H265Codec {
+    type CodecSpecificDecodeCapabilities<'a> = vk::VideoDecodeH265CapabilitiesKHR<'a>;
+    type CodecSpecificEncodeCapabilities<'a> = vk::VideoEncodeH265CapabilitiesKHR<'a>;
     type CodecSpecificEncodeQualityLevelProperties<'a> =
-        vk::VideoEncodeH264QualityLevelPropertiesKHR<'a>;
-    type NativeEncodeCodecCapabilities = NativeEncodeH264Capabilities;
+        vk::VideoEncodeH265QualityLevelPropertiesKHR<'a>;
+    type NativeEncodeCodecCapabilities = NativeEncodeH265Capabilities;
 
     fn static_decode_capabilities<'a>(
         codec_caps: &Self::CodecSpecificDecodeCapabilities<'a>,
     ) -> Self::CodecSpecificDecodeCapabilities<'static> {
-        vk::VideoDecodeH264CapabilitiesKHR {
+        vk::VideoDecodeH265CapabilitiesKHR {
             p_next: null_mut(),
             _marker: Default::default(),
             ..*codec_caps
@@ -99,7 +110,7 @@ impl CodecCapabilities for H264Codec {
     fn static_encode_capabilities<'a>(
         codec_caps: &Self::CodecSpecificEncodeCapabilities<'a>,
     ) -> Self::CodecSpecificEncodeCapabilities<'static> {
-        vk::VideoEncodeH264CapabilitiesKHR {
+        vk::VideoEncodeH265CapabilitiesKHR {
             p_next: null_mut(),
             _marker: Default::default(),
             ..*codec_caps
@@ -109,7 +120,7 @@ impl CodecCapabilities for H264Codec {
     fn static_encode_qlp<'a>(
         codec_qlp: &Self::CodecSpecificEncodeQualityLevelProperties<'a>,
     ) -> Self::CodecSpecificEncodeQualityLevelProperties<'static> {
-        vk::VideoEncodeH264QualityLevelPropertiesKHR {
+        vk::VideoEncodeH265QualityLevelPropertiesKHR {
             p_next: null_mut(),
             _marker: Default::default(),
             ..*codec_qlp
@@ -117,28 +128,27 @@ impl CodecCapabilities for H264Codec {
     }
 
     fn encode_codec_capabilities(
-        capabilities: &crate::device::caps::NativeEncodeCapabilities,
+        capabilities: &NativeEncodeCapabilities,
     ) -> Option<&Self::NativeEncodeCodecCapabilities> {
-        capabilities.h264.as_ref()
+        capabilities.h265.as_ref()
     }
 }
 
-impl<'a> CodecSpecificDecodeCapabilities for vk::VideoDecodeH264CapabilitiesKHR<'a> {}
-impl<'a> CodecSpecificEncodeCapabilities for vk::VideoEncodeH264CapabilitiesKHR<'a> {}
+impl<'a> CodecSpecificDecodeCapabilities for vk::VideoDecodeH265CapabilitiesKHR<'a> {}
+impl<'a> CodecSpecificEncodeCapabilities for vk::VideoEncodeH265CapabilitiesKHR<'a> {}
 impl<'a> CodecSpecificEncoderQualityLevelProperties
-    for vk::VideoEncodeH264QualityLevelPropertiesKHR<'a>
+    for vk::VideoEncodeH265QualityLevelPropertiesKHR<'a>
 {
     fn zeroed(&self) -> bool {
         self.preferred_rate_control_flags.as_raw() == 0
             && self.preferred_gop_frame_count == 0
             && self.preferred_idr_period == 0
             && self.preferred_consecutive_b_frame_count == 0
-            && self.preferred_temporal_layer_count == 0
+            && self.preferred_sub_layer_count == 0
             && self.preferred_constant_qp.qp_i == 0
             && self.preferred_constant_qp.qp_p == 0
             && self.preferred_constant_qp.qp_b == 0
             && self.preferred_max_l0_reference_count == 0
             && self.preferred_max_l1_reference_count == 0
-            && self.preferred_std_entropy_coding_mode_flag == 0
     }
 }
