+++ gpu-video/src/backends/vulkan/wrappers/command.rs
@@ -6,19 +6,21 @@ use std::{
 use ash::vk::{self, Handle};
 use rustc_hash::FxHashMap;
 
-use crate::{
-    VulkanCommonError,
-    device::VideoDevice,
+use crate::backends::vulkan::{
+    VulkanCommonError, VulkanDevice,
     wrappers::{ImageKey, ImageLayoutTracker, SemaphoreWaitValue},
 };
 
 struct CommandPool {
     command_pool: vk::CommandPool,
-    device: Arc<VideoDevice>,
+    device: Arc<VulkanDevice>,
 }
 
 impl CommandPool {
-    fn new(device: Arc<VideoDevice>, queue_family_index: usize) -> Result<Self, VulkanCommonError> {
+    fn new(
+        device: Arc<VulkanDevice>,
+        queue_family_index: usize,
+    ) -> Result<Self, VulkanCommonError> {
         let create_info = vk::CommandPoolCreateInfo::default()
             .flags(vk::CommandPoolCreateFlags::RESET_COMMAND_BUFFER)
             .queue_family_index(queue_family_index as u32);
@@ -80,7 +82,7 @@ pub(crate) struct CommandBufferPoolInner {
 
 impl CommandBufferPool {
     pub(crate) fn new(
-        device: Arc<VideoDevice>,
+        device: Arc<VulkanDevice>,
         queue_family_index: usize,
     ) -> Result<Self, VulkanCommonError> {
         let command_pool = CommandPool::new(device, queue_family_index)?;
