+++ gpu-video/src/backends/vulkan/wrappers/sync.rs
@@ -6,7 +6,7 @@ use std::{
 use ash::vk;
 use rustc_hash::FxHashMap;
 
-use crate::{VulkanCommonError, wrappers::ImageKey};
+use crate::backends::vulkan::{VulkanCommonError, wrappers::ImageKey};
 
 use super::Device;
