+++ gpu-video/src/lib.rs
@@ -5,40 +5,38 @@ pub mod backends;
 #[cfg(all(vulkan, not(feature = "expose-backends")))]
 pub(crate) mod backends;
 
-// TODO: After caps refactor cfg for instance and adapter won't be needed
+#[cfg(feature = "expose-parsers")]
+pub mod parser;
+#[cfg(not(feature = "expose-parsers"))]
+pub(crate) mod parser;
+
+// TODO: The modules below should compile on macos
 #[cfg(vulkan)]
 mod adapter;
 #[cfg(vulkan)]
-mod instance;
-
-#[cfg(vulkan)]
-pub(crate) mod codec;
+pub mod capabilities;
 #[cfg(vulkan)]
 mod device;
-// TODO: cfg(vulkan) will not be needed once we add proper metal support
 #[cfg(all(vulkan, feature = "wgpu"))]
 mod global_registry;
+#[cfg(vulkan)]
+mod instance;
+#[cfg(all(vulkan, feature = "wgpu"))]
+pub(crate) mod wgpu_helpers;
+
+// TODO: The modules below should be made backend agnostic
 #[cfg(vulkan)]
 mod vulkan_decoder;
 #[cfg(vulkan)]
 mod vulkan_encoder;
 #[cfg(all(vulkan, feature = "transcoder"))]
 mod vulkan_transcoder;
-#[cfg(all(vulkan, feature = "wgpu"))]
-pub(crate) mod wgpu_helpers;
-#[cfg(vulkan)]
-pub(crate) mod wrappers;
 
-// TODO: This could be inlined with lib.rs when metal support is added
+// TODO: Rename to prelude? Or exports?
 #[cfg(vulkan)]
 mod vulkan_video;
 #[cfg(vulkan)]
 pub use vulkan_video::*;
-
-#[cfg(feature = "expose-parsers")]
-pub mod parser;
-#[cfg(not(feature = "expose-parsers"))]
-pub(crate) mod parser;
 
 // If vulkan is unsupported and parsers are not exposed
 #[cfg(not(any(vulkan, feature = "expose-parsers")))]
