+++ gpu-video/src/vulkan_transcoder.rs
@@ -4,9 +4,12 @@ use ash::vk;
 
 use crate::{
     EncodedInputChunk, EncodedOutputChunk, OutputFrame, VideoDecoderError, VideoEncoderError,
-    VulkanCommonError,
-    codec::{EncodeCodec, h264::H264Codec, h265::H265Codec},
-    device::{EncoderOutputParameters, Rational, VideoDevice},
+    backends::vulkan::{
+        VulkanCommonError, VulkanDevice,
+        codec::{EncodeCodec, h264::H264Codec, h265::H265Codec},
+        wrappers::{DecodeInputBuffer, DecodingQueryPool, SemaphoreWaitValue},
+    },
+    device::{EncoderOutputParameters, Rational},
     parameters::{H264Profile, H265Profile, ScalingAlgorithm},
     parser::{
         decoder_instructions::{DecoderInstruction, compile_to_decoder_instructions},
@@ -18,7 +21,6 @@ use crate::{
     },
     vulkan_encoder::{Encoder, FullEncoderParameters, VulkanEncoder},
     vulkan_transcoder::pipeline::{OutputConfig, ResizeSubmission, ResizingPipeline},
-    wrappers::{DecodeInputBuffer, DecodingQueryPool, SemaphoreWaitValue},
 };
 
 mod pipeline;
@@ -78,7 +80,7 @@ pub(crate) struct ResizedImages {
 }
 
 pub struct Transcoder {
-    device: Arc<VideoDevice>,
+    device: Arc<VulkanDevice>,
     decoder: VulkanDecoder<'static>,
     parser: H264Parser,
     reference_ctx: ReferenceContext,
@@ -89,7 +91,7 @@ pub struct Transcoder {
 
 impl Transcoder {
     pub(crate) fn new(
-        device: Arc<VideoDevice>,
+        device: Arc<VulkanDevice>,
         config: TranscoderParameters,
     ) -> Result<Self, VideoTranscoderError> {
         let decoder = VulkanDecoder::new(
